+++ sdk/resourcemanager/containerservicefleet/armcontainerservicefleet/client_factory.go
@@ -17,8 +17,7 @@ import (
 // Don't use this type directly, use NewClientFactory instead.
 type ClientFactory struct {
 	subscriptionID string
-	credential     azcore.TokenCredential
-	options        *arm.ClientOptions
+	internal       *arm.Client
 }
 
 // NewClientFactory creates a new instance of ClientFactory with the specified values.
@@ -27,42 +26,51 @@ type ClientFactory struct {
 //   - credential - used to authorize requests. Usually a credential from azidentity.
 //   - options - pass nil to accept the default values.
 func NewClientFactory(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*ClientFactory, error) {
-	_, err := arm.NewClient(moduleName, moduleVersion, credential, options)
+	internal, err := arm.NewClient(moduleName, moduleVersion, credential, options)
 	if err != nil {
 		return nil, err
 	}
 	return &ClientFactory{
-		subscriptionID: subscriptionID, credential: credential,
-		options: options.Clone(),
+		subscriptionID: subscriptionID,
+		internal:       internal,
 	}, nil
 }
 
 // NewFleetMembersClient creates a new instance of FleetMembersClient.
 func (c *ClientFactory) NewFleetMembersClient() *FleetMembersClient {
-	subClient, _ := NewFleetMembersClient(c.subscriptionID, c.credential, c.options)
-	return subClient
+	return &FleetMembersClient{
+		subscriptionID: c.subscriptionID,
+		internal:       c.internal,
+	}
 }
 
 // NewFleetUpdateStrategiesClient creates a new instance of FleetUpdateStrategiesClient.
 func (c *ClientFactory) NewFleetUpdateStrategiesClient() *FleetUpdateStrategiesClient {
-	subClient, _ := NewFleetUpdateStrategiesClient(c.subscriptionID, c.credential, c.options)
-	return subClient
+	return &FleetUpdateStrategiesClient{
+		subscriptionID: c.subscriptionID,
+		internal:       c.internal,
+	}
 }
 
 // NewFleetsClient creates a new instance of FleetsClient.
 func (c *ClientFactory) NewFleetsClient() *FleetsClient {
-	subClient, _ := NewFleetsClient(c.subscriptionID, c.credential, c.options)
-	return subClient
+	return &FleetsClient{
+		subscriptionID: c.subscriptionID,
+		internal:       c.internal,
+	}
 }
 
 // NewOperationsClient creates a new instance of OperationsClient.
 func (c *ClientFactory) NewOperationsClient() *OperationsClient {
-	subClient, _ := NewOperationsClient(c.credential, c.options)
-	return subClient
+	return &OperationsClient{
+		internal: c.internal,
+	}
 }
 
 // NewUpdateRunsClient creates a new instance of UpdateRunsClient.
 func (c *ClientFactory) NewUpdateRunsClient() *UpdateRunsClient {
-	subClient, _ := NewUpdateRunsClient(c.subscriptionID, c.credential, c.options)
-	return subClient
+	return &UpdateRunsClient{
+		subscriptionID: c.subscriptionID,
+		internal:       c.internal,
+	}
 }