+++ sdk/resourcemanager/containerservicefleet/armcontainerservicefleet/fleets_client_example_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservicefleet/armcontainerservicefleet"
 )
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/678aeba91061a67e9af8b9aaf41f85b8aea06c7c/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/stable/2023-10-15/examples/Fleets_ListBySub.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/92de53a5f1e0e03c94b40475d2135d97148ed014/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/preview/2024-02-02-preview/examples/Fleets_ListBySub.json
 func ExampleFleetsClient_NewListBySubscriptionPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -61,6 +61,15 @@ func ExampleFleetsClient_NewListBySubscriptionPager() {
 		// 			},
 		// 			ETag: to.Ptr("23ujdflewrj3="),
 		// 			Properties: &armcontainerservicefleet.FleetProperties{
+		// 				HubProfile: &armcontainerservicefleet.FleetHubProfile{
+		// 					AgentProfile: &armcontainerservicefleet.AgentProfile{
+		// 						VMSize: to.Ptr("Standard_DS1"),
+		// 					},
+		// 					DNSPrefix: to.Ptr("dnsprefix1"),
+		// 					Fqdn: to.Ptr("dnsprefix1-abcd1234.flt.eastus.azmk8s.io"),
+		// 					KubernetesVersion: to.Ptr("1.22.4"),
+		// 					PortalFqdn: to.Ptr("dnsprefix1-abcd1234.flt.eastus.azmk8s.io"),
+		// 				},
 		// 				ProvisioningState: to.Ptr(armcontainerservicefleet.FleetProvisioningStateSucceeded),
 		// 			},
 		// 	}},
@@ -68,7 +77,7 @@ func ExampleFleetsClient_NewListBySubscriptionPager() {
 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/678aeba91061a67e9af8b9aaf41f85b8aea06c7c/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/stable/2023-10-15/examples/Fleets_ListByResourceGroup.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/92de53a5f1e0e03c94b40475d2135d97148ed014/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/preview/2024-02-02-preview/examples/Fleets_ListByResourceGroup.json
 func ExampleFleetsClient_NewListByResourceGroupPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -111,6 +120,15 @@ func ExampleFleetsClient_NewListByResourceGroupPager() {
 		// 			},
 		// 			ETag: to.Ptr("23ujdflewrj3="),
 		// 			Properties: &armcontainerservicefleet.FleetProperties{
+		// 				HubProfile: &armcontainerservicefleet.FleetHubProfile{
+		// 					AgentProfile: &armcontainerservicefleet.AgentProfile{
+		// 						VMSize: to.Ptr("Standard_DS1"),
+		// 					},
+		// 					DNSPrefix: to.Ptr("dnsprefix1"),
+		// 					Fqdn: to.Ptr("dnsprefix1-abcd1234.flt.eastus.azmk8s.io"),
+		// 					KubernetesVersion: to.Ptr("1.22.4"),
+		// 					PortalFqdn: to.Ptr("dnsprefix1-abcd1234.flt.eastus.azmk8s.io"),
+		// 				},
 		// 				ProvisioningState: to.Ptr(armcontainerservicefleet.FleetProvisioningStateSucceeded),
 		// 			},
 		// 	}},
@@ -118,7 +136,7 @@ func ExampleFleetsClient_NewListByResourceGroupPager() {
 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/678aeba91061a67e9af8b9aaf41f85b8aea06c7c/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/stable/2023-10-15/examples/Fleets_Get.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/92de53a5f1e0e03c94b40475d2135d97148ed014/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/preview/2024-02-02-preview/examples/Fleets_Get.json
 func ExampleFleetsClient_Get() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -155,12 +173,21 @@ func ExampleFleetsClient_Get() {
 	// 	},
 	// 	ETag: to.Ptr("23ujdflewrj3="),
 	// 	Properties: &armcontainerservicefleet.FleetProperties{
+	// 		HubProfile: &armcontainerservicefleet.FleetHubProfile{
+	// 			AgentProfile: &armcontainerservicefleet.AgentProfile{
+	// 				VMSize: to.Ptr("Standard_DS1"),
+	// 			},
+	// 			DNSPrefix: to.Ptr("dnsprefix1"),
+	// 			Fqdn: to.Ptr("dnsprefix1-abcd1234.flt.eastus.azmk8s.io"),
+	// 			KubernetesVersion: to.Ptr("1.22.4"),
+	// 			PortalFqdn: to.Ptr("dnsprefix1-abcd1234.flt.eastus.azmk8s.io"),
+	// 		},
 	// 		ProvisioningState: to.Ptr(armcontainerservicefleet.FleetProvisioningStateSucceeded),
 	// 	},
 	// }
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/678aeba91061a67e9af8b9aaf41f85b8aea06c7c/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/stable/2023-10-15/examples/Fleets_CreateOrUpdate.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/92de53a5f1e0e03c94b40475d2135d97148ed014/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/preview/2024-02-02-preview/examples/Fleets_CreateOrUpdate.json
 func ExampleFleetsClient_BeginCreateOrUpdate() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -177,7 +204,14 @@ func ExampleFleetsClient_BeginCreateOrUpdate() {
 			"archv2": to.Ptr(""),
 			"tier":   to.Ptr("production"),
 		},
-		Properties: &armcontainerservicefleet.FleetProperties{},
+		Properties: &armcontainerservicefleet.FleetProperties{
+			HubProfile: &armcontainerservicefleet.FleetHubProfile{
+				AgentProfile: &armcontainerservicefleet.AgentProfile{
+					VMSize: to.Ptr("Standard_DS1"),
+				},
+				DNSPrefix: to.Ptr("dnsprefix1"),
+			},
+		},
 	}, &armcontainerservicefleet.FleetsClientBeginCreateOrUpdateOptions{IfMatch: nil,
 		IfNoneMatch: nil,
 	})
@@ -210,12 +244,21 @@ func ExampleFleetsClient_BeginCreateOrUpdate() {
 	// 	},
 	// 	ETag: to.Ptr("23ujdflewrj3="),
 	// 	Properties: &armcontainerservicefleet.FleetProperties{
+	// 		HubProfile: &armcontainerservicefleet.FleetHubProfile{
+	// 			AgentProfile: &armcontainerservicefleet.AgentProfile{
+	// 				VMSize: to.Ptr("Standard_DS1"),
+	// 			},
+	// 			DNSPrefix: to.Ptr("dnsprefix1"),
+	// 			Fqdn: to.Ptr("dnsprefix1-abcd1234.flt.eastus.azmk8s.io"),
+	// 			KubernetesVersion: to.Ptr("1.22.4"),
+	// 			PortalFqdn: to.Ptr("dnsprefix1-abcd1234.flt.eastus.azmk8s.io"),
+	// 		},
 	// 		ProvisioningState: to.Ptr(armcontainerservicefleet.FleetProvisioningStateSucceeded),
 	// 	},
 	// }
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/678aeba91061a67e9af8b9aaf41f85b8aea06c7c/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/stable/2023-10-15/examples/Fleets_PatchTags.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/92de53a5f1e0e03c94b40475d2135d97148ed014/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/preview/2024-02-02-preview/examples/Fleets_PatchTags.json
 func ExampleFleetsClient_BeginUpdate() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -261,12 +304,21 @@ func ExampleFleetsClient_BeginUpdate() {
 	// 	},
 	// 	ETag: to.Ptr("23ujdflewrj3="),
 	// 	Properties: &armcontainerservicefleet.FleetProperties{
+	// 		HubProfile: &armcontainerservicefleet.FleetHubProfile{
+	// 			AgentProfile: &armcontainerservicefleet.AgentProfile{
+	// 				VMSize: to.Ptr("Standard_DS1"),
+	// 			},
+	// 			DNSPrefix: to.Ptr("dnsprefix1"),
+	// 			Fqdn: to.Ptr("dnsprefix1-abcd1234.flt.eastus.azmk8s.io"),
+	// 			KubernetesVersion: to.Ptr("1.22.4"),
+	// 			PortalFqdn: to.Ptr("dnsprefix1-abcd1234.flt.eastus.azmk8s.io"),
+	// 		},
 	// 		ProvisioningState: to.Ptr(armcontainerservicefleet.FleetProvisioningStateSucceeded),
 	// 	},
 	// }
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/678aeba91061a67e9af8b9aaf41f85b8aea06c7c/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/stable/2023-10-15/examples/Fleets_Delete.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/92de53a5f1e0e03c94b40475d2135d97148ed014/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/preview/2024-02-02-preview/examples/Fleets_Delete.json
 func ExampleFleetsClient_BeginDelete() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -287,7 +339,7 @@ func ExampleFleetsClient_BeginDelete() {
 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/678aeba91061a67e9af8b9aaf41f85b8aea06c7c/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/stable/2023-10-15/examples/Fleets_ListCredentialsResult.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/92de53a5f1e0e03c94b40475d2135d97148ed014/specification/containerservice/resource-manager/Microsoft.ContainerService/fleet/preview/2024-02-02-preview/examples/Fleets_ListCredentialsResult.json
 func ExampleFleetsClient_ListCredentials() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {