+++ sdk/storage/azfile/file/client.go
@@ -10,6 +10,12 @@ import (
 	"bytes"
 	"context"
 	"errors"
+	"io"
+	"os"
+	"strings"
+	"sync"
+	"time"
+
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
@@ -21,11 +27,6 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azfile/internal/generated"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azfile/internal/shared"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azfile/sas"
-	"io"
-	"os"
-	"strings"
-	"sync"
-	"time"
 )
 
 // ClientOptions contains the optional parameters when creating a Client.
@@ -42,8 +43,10 @@ type Client base.Client[generated.FileClient]
 // Note that ClientOptions.FileRequestIntent is currently required for token authentication.
 func NewClient(fileURL string, cred azcore.TokenCredential, options *ClientOptions) (*Client, error) {
 	audience := base.GetAudience((*base.ClientOptions)(options))
-	authPolicy := runtime.NewBearerTokenPolicy(cred, []string{audience}, nil)
 	conOptions := shared.GetClientOptions(options)
+	authPolicy := runtime.NewBearerTokenPolicy(cred, []string{audience}, &policy.BearerTokenOptions{
+		InsecureAllowCredentialWithHTTP: conOptions.InsecureAllowCredentialWithHTTP,
+	})
 	plOpts := runtime.PipelineOptions{
 		PerRetry: []policy.Policy{authPolicy},
 	}
@@ -338,7 +341,6 @@ func (f *Client) GetSASURL(permissions sas.FilePermissions, expiry time.Time, o 
 
 	qps, err := sas.SignatureValues{
 		Version:     sas.Version,
-		Protocol:    sas.ProtocolHTTPS,
 		ShareName:   urlParts.ShareName,
 		FilePath:    urlParts.DirectoryOrFilePath,
 		Permissions: permissions.String(),